python/requirements.txt
# FastAPI and server
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6

# API and network
aiohttp>=3.11.13,<4.0.0
requests==2.28.2
websockets>=11.0.0,<15.0.0

# System monitoring (Windows only)
psutil==5.9.4
pywin32==306

# Screenshot capture
mss==7.0.1
Pillow>=10.1.0,<11.0.0

# Utils
pydantic>=1.10.6,<2.0.0
python-dotenv==1.0.0
typing-extensions>=4.12.2,<5.0.0

# Authentication & Supabase
PyJWT==2.6.0
email-validator==2.0.0
supabase==2.14.0  # Latest Python client for Supabase
